requirements.txt
python-osc>=1.8.0
pyyaml>=6.0.0
irctokens>=2.0.0
